update lock position
